Add Deriving Script

+++ SuperMarket.API/Connection/PatientFunction.cs
@@ -56,7 +56,7 @@ namespace SuperMarket.API.PatientFunction
 
         }
 
-        public static bool DeletePatient(Patient value)
+        public static void DeletePatient(Patient value)
         {
             Console.WriteLine("Show Patient start");
             var connString = "mongodb://127.0.0.1:27017";
@@ -66,43 +66,24 @@ namespace SuperMarket.API.PatientFunction
 
             Console.WriteLine("Delte Patient Start");
 
-            var firstFilter = Builders<Patient>.Filter.Eq("idDoctor", value.idDoctor);
+            var firstFilter = Builders<Patient>.Filter.Eq("idDoctor", value._id);
 
-            var cursor = collection.Find(firstFilter).ToCursor();
-            foreach (var collect in cursor.ToEnumerable())
+            if (firstFilter != null)
             {
-                // var firstNameFilter = Builders<Patient>.Filter.Eq("idDoctor", value.idDoctor);
-                // var cursorFirst = collection.Find(firstNameFilter).ToCursor();
-                // foreach (var first in cursorFirst.ToEnumerable())
-                // { 
-                //     var cursorFirst = collection.Find(firstNameFilter).ToCursor();
-                // foreach (var first in cursorFirst.ToEnumerable())
-                // { 
-
-                // }
-                // }
-
-                // var lastNameFilter = Builders<Patient>.Filter.Eq("idDoctor", value.idDoctor);
-                // if (collect.firstNamePatient.Equals(value.firstNamePatient))
-                // {
-                //     if (collect.lastNamePatient.Equals(value.lastNamePatient))
-                //     {
-                //         Console.WriteLine(" patient deleted");
-                //         collection.DeleteOne(firstFilter);
-                //         return true;
-                //     }
-                // }
+                collection.DeleteOne(firstFilter);
+                Console.WriteLine(" deleted ");
             }
 
 
-            var highExamScoreFilter = Builders<Patient>.Filter.ElemMatch<BsonValue>(
-            "scores", new BsonDocument { { "type", "exam" },
-            { "score", new BsonDocument { { "$gte", 95 } } }
-            });
+            // var cursor = collection.Find(firstFilter).ToCursor();
+            // foreach (var collect in cursor.ToEnumerable())
+            // {
+            //     Console.WriteLine();
+            // }
 
-            collection.DeleteOne(firstFilter);
+
+            // collection.DeleteOne(firstFilter);
             Console.WriteLine(" nothing to delete");
-            return false;
 
         }
